+++ src/com/codeup/movies/Category.java
@@ -4,15 +4,15 @@
 package com.codeup.movies;
 
 public class Category {
-    private final int id;
+    private final long id;
     private final String name;
 
-    public Category(int id, String name) {
+    public Category(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int id() {
+    public long id() {
         return id;
     }
 
@@ -21,6 +21,6 @@ public class Category {
     }
 
     public static Category named(String category) {
-        return new Category(0, category);
+        return new Category(0L, category);
     }
 }
